Search count of literature

This function counts the total number of search results from Scopus, Web of Science, or BASE. A check will be conducted via gnr_resolve to validate the genus and species names.

Count(db, search, genus, species = NULL, synonyms, additionalkeywords)

Arguments

db

Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").

search

Search fields. Title only ("t") or title, abstract, or keywords ("tak").

genus

Genus classification from the binomial name.

species

Species classification from the binomial name.

synonyms

Alternate species names; optional.

additionalkeywords

Optional search terms.

Value

Search count of the genus or species with the given genus and/or species.

Examples

if (FALSE) {
Count(db = "scopus",
      search = "t",
      genus = "Osphranter",
      species = "rufus")
}
if (FALSE) {
Count(db = "scopus",
      search = "t",
      genus = "Osphranter",
      species = "rufus",
      synonyms = "Macropus rufus",
      additionalkeywords = "conserv*")
}