Extract year

Extracts the year of each publication of the output from any of the Fetch functions and counts the number of publications each year.

getYear(data, genus, species)

Arguments

data

Output from any of the fetch function.

genus

Genus classification from the binomial name.

species

Species classification from the binomial name.

Value

A dataframe with the year and frequency of the publications

Examples

if (FALSE) {
Woylie <- Fetch(db = "scopus",
                search = "tak",
                genus = "Bettongia",
                species = "penicillata")
}
getYear(data = Woylie,
        genus = "Bettongia",
        species = "penicillata")
#> # A tibble: 44 × 3
#>     Year  Freq spp                  
#>    <dbl> <int> <chr>                
#>  1  1977     1 Bettongia penicillata
#>  2  1978     0 Bettongia penicillata
#>  3  1979     1 Bettongia penicillata
#>  4  1980     1 Bettongia penicillata
#>  5  1981     0 Bettongia penicillata
#>  6  1982     0 Bettongia penicillata
#>  7  1983     0 Bettongia penicillata
#>  8  1984     0 Bettongia penicillata
#>  9  1985     2 Bettongia penicillata
#> 10  1986     1 Bettongia penicillata
#> # … with 34 more rows